refactor: MyPageElements 추가
